+++ src/kimi_cli/tools/shell/__init__.py
@@ -18,7 +18,7 @@ from kimi_cli.tools.display import BackgroundTaskDisplayBlock, ShellDisplayBlock
 from kimi_cli.tools.file.rg_path import find_existing_rg, format_rg_command
 from kimi_cli.tools.utils import ToolRejectedError, ToolResultBuilder, load_desc
 from kimi_cli.utils.environment import Environment
-from kimi_cli.utils.subprocess_env import get_clean_env
+from kimi_cli.utils.subprocess_env import get_noninteractive_env
 from kimi_cli.wire.types import ToolCallOutput
 
 MAX_FOREGROUND_TIMEOUT = 5 * 60
@@ -255,7 +255,11 @@ class Shell(CallableTool2[Params]):
                 else:
                     break
 
-        process = await kaos.exec(*self._shell_args(command), env=get_clean_env())
+        process = await kaos.exec(*self._shell_args(command), env=get_noninteractive_env())
+
+        # Close stdin immediately so interactive prompts (e.g. git password) get
+        # EOF instead of hanging forever waiting for input that will never come.
+        process.stdin.close()
 
         try:
             await asyncio.wait_for(